+++ contrib/ros/robots/sawyer.py
@@ -100,7 +100,7 @@ class Sawyer(Robot):
 
     @property
     def observation_space(self):
-        return gym.spaces.Box(-np.inf, np.inf, shape=self.get_observation().shape)
+        return gym.spaces.Box(-np.inf, np.inf, shape=self.get_observation().shape, dtype=np.float32)
 
     def send_command(self, commands):
         """
@@ -164,4 +164,4 @@ class Sawyer(Robot):
                     'Control mode %s is not known!' % self._control_mode)
         return gym.spaces.Box(
             np.concatenate((lower_bounds, np.array([0]))),
-            np.concatenate((upper_bounds, np.array([100]))))
+            np.concatenate((upper_bounds, np.array([100]))), dtype=np.float32)
