+++ rllab/envs/mujoco/hill/walker2d_hill_env.py
@@ -13,4 +13,4 @@ class Walker2DHillEnv(HillEnv):
     @overrides
     def _mod_hfield(self, hfield):
         # clear a flat patch for the robot to start off from
-        return terrain.clear_patch(hfield, gym.spaces.Box(np.array([-2.0, -2.0]), np.array([-0.5, -0.5])))
+        return terrain.clear_patch(hfield, gym.spaces.Box(np.array([-2.0, -2.0]), np.array([-0.5, -0.5])), dtype=np.float32)
